# public/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Launchpad API",
    "description": "API documentation for the Launchpad application",
    "version": "1.0.0",
    "contact": {
      "email": "support@launchpad.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Local server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique location identifier",
            "example": "loc_abc123"
          },
          "name": {
            "type": "string",
            "description": "Location name",
            "example": "Main Warehouse"
          },
          "code": {
            "type": "string",
            "description": "Location code for easy identification and sorting",
            "example": "WH-MAIN"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Detailed location description",
            "example": "Primary storage facility for all products and components"
          },
          "path": {
            "type": "string",
            "description": "Hierarchical path showing the location's position in the tree structure",
            "example": "/loc_abc123/loc_def456/"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the location is currently active and available for use",
            "example": true
          },
          "parent_id": {
            "type": "string",
            "nullable": true,
            "description": "Parent location ID (null for root/top-level locations)",
            "example": "loc_parent123"
          },
          "organization": {
            "type": "object",
            "description": "Organization this location belongs to",
            "properties": {
              "id": {
                "type": "string",
                "example": "org_xyz789"
              },
              "name": {
                "type": "string",
                "example": "ACME Corporation"
              }
            }
          },
          "parent": {
            "type": "object",
            "nullable": true,
            "description": "Parent location details (null for root locations)",
            "properties": {
              "id": {
                "type": "string",
                "example": "loc_parent123"
              },
              "name": {
                "type": "string",
                "example": "Main Building"
              },
              "code": {
                "type": "string",
                "example": "MB-01"
              },
              "path": {
                "type": "string",
                "example": "/loc_parent123/"
              }
            }
          },
          "children": {
            "type": "array",
            "description": "Direct child locations (first level only)",
            "items": {
              "$ref": "#/components/schemas/LocationSummary"
            }
          },
          "childrenRecursive": {
            "type": "array",
            "description": "All nested child locations recursively (complete hierarchy)",
            "items": {
              "$ref": "#/components/schemas/Location"
            }
          },
          "items": {
            "type": "array",
            "description": "Items stored in this location (when included with ?with=items)",
            "items": {
              "$ref": "#/components/schemas/ItemSummary"
            }
          },
          "items_count": {
            "type": "integer",
            "description": "Total number of items stored in this location",
            "example": 45
          },
          "attachments": {
            "type": "array",
            "description": "Files and documents attached to this location",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Location creation timestamp",
            "example": "2024-01-15T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last modification timestamp",
            "example": "2024-01-15T10:30:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "path",
          "is_active",
          "created_at",
          "updated_at"
        ]
      },
...
      "LocationCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Location name (must be unique within the organization)",
            "minLength": 1,
            "maxLength": 255,
            "example": "Main Warehouse"
          },
          "code": {
            "type": "string",
            "description": "Location code for identification (must be unique within the organization)",
            "minLength": 1,
            "maxLength": 50,
            "pattern": "^[A-Z0-9_-]+$",
            "example": "WH-MAIN"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Optional detailed description of the location",
            "maxLength": 1000,
            "example": "Primary storage facility for all products and electronic components"
          },
          "parent_id": {
            "type": "string",
            "nullable": true,
            "description": "Parent location ID (null for root/top-level locations). Must be an existing location ID.",
            "example": "loc_parent123"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the location should be active upon creation",
            "default": true,
            "example": true
          }
        },
        "required": [
          "name",
          "code"
        ],
        "example": {
          "name": "Electronics Section",
          "code": "WH-MAIN-ELEC",
          "description": "Dedicated area for electronic components and devices",
          "parent_id": "loc_abc123",
          "is_active": true
        }
      },
...
      "LocationSummary": {
        "type": "object",
        "description": "Simplified location information for references and listings",
        "properties": {
          "id": {
            "type": "string",
            "description": "Location identifier",
            "example": "loc_abc123"
          },
          "name": {
            "type": "string",
            "description": "Location name",
            "example": "Main Warehouse"
          },
          "code": {
            "type": "string",
            "description": "Location code",
            "example": "WH-MAIN"
          },
          "path": {
            "type": "string",
            "description": "Hierarchical path",
            "example": "/loc_abc123/"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the location is active",
            "example": true
          },
          "items_count": {
            "type": "integer",
            "description": "Number of items in this location",
            "example": 25
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "path",
          "is_active"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/locations": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Get all locations",
        "description": "Retrieve all locations with advanced filtering and optional hierarchical view. Supports searching by name, code, description, active status, and parent relationships. By default returns hierarchical structure with nested children.",
        "operationId": "getLocations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "hierarchy",
            "in": "query",
            "description": "Return locations in hierarchical structure (true) or flat list (false). Hierarchical view shows nested parent-child relationships.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            },
            "example": true
          },
          {
            "name": "name",
            "in": "query",
            "description": "Filter by location name (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Warehouse"
          },
          {
            "name": "code",
            "in": "query",
            "description": "Filter by location code (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "WH-"
          },
          {
            "name": "description",
            "in": "query",
            "description": "Filter by location description (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "storage"
          },
          {
            "name": "is_active",
            "in": "query",
            "description": "Filter by active status. Use 'true' for active, 'false' for inactive locations.",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "example": true
          },
          {
            "name": "parent_id",
            "in": "query",
            "description": "Filter by parent location ID. Use 'null' to get root locations (no parent).",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "loc_abc123"
          },
          {
            "name": "with",
            "in": "query",
            "description": "Include related data (comma-separated). Available: parent, children, childrenRecursive, organization, items, attachments",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "parent,children,organization,items"
          }
        ],
        "responses": {
          "200": {
            "description": "Locations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Location"
                      }
                    }
                  }
                },
                "examples": {
                  "hierarchical": {
                    "summary": "Hierarchical locations with nested children",
                    "value": {
                      "status": "success",
                      "message": "Success",
                      "data": [
                        {
                          "id": "loc_abc123",
                          "name": "Main Warehouse",
                          "code": "WH-MAIN",
                          "description": "Primary storage facility for all products",
                          "path": "/loc_abc123/",
                          "is_active": true,
                          "parent_id": null,
                          "organization": {
                            "id": "org_xyz789",
                            "name": "ACME Corporation"
                          },
                          "childrenRecursive": [
                            {
                              "id": "loc_def456",
                              "name": "Electronics Section",
                              "code": "WH-MAIN-ELEC",
                              "description": "Dedicated area for electronic components",
                              "path": "/loc_abc123/loc_def456/",
                              "is_active": true,
                              "parent_id": "loc_abc123",
                              "childrenRecursive": [
                                {
                                  "id": "loc_ghi789",
                                  "name": "Shelf A1",
                                  "code": "WH-MAIN-ELEC-A1",
                                  "description": "Top shelf for small electronics",
                                  "path": "/loc_abc123/loc_def456/loc_ghi789/",
                                  "is_active": true,
                                  "parent_id": "loc_def456",
                                  "childrenRecursive": [],
                                  "items_count": 25
                                }
                              ],
                              "items_count": 45
                            }
                          ],
                          "items_count": 120,
                          "created_at": "2024-01-15T10:30:00Z",
                          "updated_at": "2024-01-15T10:30:00Z"
                        }
                      ]
                    }
                  },
                  "filtered": {
                    "summary": "Filtered locations by active status",
                    "value": {
                      "status": "success",
                      "message": "Success",
                      "data": [
                        {
                          "id": "loc_abc123",
                          "name": "Main Warehouse",
                          "code": "WH-MAIN",
                          "description": "Primary storage facility",
                          "path": "/loc_abc123/",
                          "is_active": true,
                          "parent_id": null,
                          "items_count": 120,
                          "created_at": "2024-01-15T10:30:00Z",
                          "updated_at": "2024-01-15T10:30:00Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unknown query parameter: invalid_param"
                    },
                    "data": {
                      "type": "null",
                      "example": null
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "message": "Unknown query parameter: invalid_param",
                  "code": 400
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "null",
                      "example": null
                    }
                  }
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User must belong to an organization or insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "User must belong to an organization"
                    },
                    "data": {
                      "type": "null",
                      "example": null
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "message": "User must belong to an organization",
                  "code": 400
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Locations"
        ],
        "summary": "Create a new location",
        "description": "Create a new location in the organization. Supports hierarchical structure by specifying parent_id. The system automatically generates the path based on the hierarchy.",
        "operationId": "createLocation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationCreateRequest"
              },
              "examples": {
                "root_location": {
                  "summary": "Root level warehouse",
                  "value": {
                    "name": "Main Warehouse",
                    "code": "WH-MAIN",
                    "description": "Primary storage facility for all products",
                    "parent_id": null,
                    "is_active": true
                  }
                },
                "child_location": {
                  "summary": "Child location within warehouse",
                  "value": {
                    "name": "Electronics Section",
                    "code": "WH-MAIN-ELEC",
                    "description": "Dedicated area for electronic components and devices",
                    "parent_id": "loc_abc123",
                    "is_active": true
                  }
                },
                "nested_location": {
                  "summary": "Deeply nested storage location",
                  "value": {
                    "name": "Shelf A1",
                    "code": "WH-MAIN-ELEC-A1",
                    "description": "Top shelf for small electronic components",
                    "parent_id": "loc_def456",
                    "is_active": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Location created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Resource created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Location"
                    }
                  }
                },
                "examples": {
                  "created_location": {
                    "summary": "Newly created location",
                    "value": {
                      "status": "success",
                      "message": "Resource created successfully",
                      "data": {
                        "id": "loc_new123",
                        "name": "Electronics Section",
                        "code": "WH-MAIN-ELEC",
                        "description": "Dedicated area for electronic components",
                        "path": "/loc_abc123/loc_new123/",
                        "is_active": true,
                        "parent_id": "loc_abc123",
                        "parent": {
                          "id": "loc_abc123",
                          "name": "Main Warehouse",
                          "code": "WH-MAIN"
                        },
                        "organization": {
                          "id": "org_xyz789",
                          "name": "ACME Corporation"
                        },
                        "items_count": 0,
                        "created_at": "2024-01-16T14:20:00Z",
                        "updated_at": "2024-01-16T14:20:00Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Empty data or invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data cannot be empty"
                    },
                    "data": {
                      "type": "null",
                      "example": null
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "message": "Data cannot be empty",
                  "code": 400
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "null",
                      "example": null
                    }
                  }
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User must belong to an organization or insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "User must belong to an organization"
                    },
                    "data": {
                      "type": "null",
                      "example": null
                    }
                  }
                },
                "example": {
                  "status": "error",
                  "message": "User must belong to an organization",
                  "code": 400
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The given data was invalid."
                    },
                    "errors": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "The location name is required and cannot be empty"
                          ]
                        },
                        "code": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "This location code already exists in your organization"
                          ]
                        },
                        "parent_id": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "The selected parent location does not exist in your organization"
                          ]
                        }
                      }
                    }
                  }
                },
                "example": {
                  "message": "The given data was invalid."
                }
              }
            }
          }
        }
      }
    },
...
  }
}